+++ app/javascript/vue/tasks/digitize/components/collectionObject/containerItems.vue
@@ -48,7 +48,6 @@ export default {
     newDigitalization () {
       this.$store.dispatch(ActionNames.NewCollectionObject)
       this.$store.dispatch(ActionNames.NewIdentifier)
-      this.$store.commit(MutationNames.NewTaxonDetermination)
       this.$store.commit(MutationNames.SetTaxonDeterminations, [])
     },